refactor(events): narrow the delta events' amount to uint120
All four delta events declared a uint256 amount next to uint120 balances, so a
consumer decoding the tuple saw two widths for three quantities of the same kind.

No value changes. In capture and refund the emit now passes the uint120 local that
ALREADY existed for the storage write (captureAmount120 / refundAmount120) instead of
the uint256 parameter — so no new cast is introduced, and the two existing safe-cast
comments still carry the justification. void and release were already emitting a
uint120 local that only the event signature widened back.

BREAKING: topic0 changes on all four.

  PaymentCaptured  0x460e11c3… → 0x77b933fc…
  PaymentVoided    0xe60803c7… → 0xb37b3ce4…
  PaymentReleased  0xaf7a769d… → 0x90204216…
  PaymentRefunded  0xaa05bfb9… → 0xe53d95a7…

So this must NOT deploy alone. It needs a VERSION bump and the coordinated rollout,
plus a full reindex. rail0#53 (the minimum refund window) is also breaking and also
unmerged — one bump covers both, and shipping them together avoids two reindexes.
VERSION is deliberately left at 1.3.0 here, as in #53: bumping it is a release
decision across six live deployments.

The companion indexer change is commercelayer/rail0-indexer#78.

Gas: +18 on capture (178,347 → 178,365), from reading the local rather than the
parameter. A uint120 still occupies a full word in event data, so nothing is saved
there either — this is consistency, not optimisation.

The four locally re-declared events in RAIL0.t.sol had to be narrowed too, and all
four event tests failed until they were. That is the clearest local evidence topic0
moved, and it is noted in the test file so the next reader sees why they exist.

Suite: 120 tests pass. forge fmt --check clean.

Closes #47

+++ contracts/test/RAIL0.t.sol
@@ -1941,11 +1941,16 @@ contract RAIL0Test is Test {
     // must mirror) rather than against a hand-written literal.
 
     /// Re-declared locally: vm.expectEmit needs the event in the test's scope.
+    ///
+    /// These doubles as a golden signature check. Narrowing `amount` in the contract
+    /// made all four of these tests fail until they were narrowed too — which is the
+    /// clearest local evidence that topic0 changed, and why this cannot ship without a
+    /// version bump and a reindex. (#47)
     event PaymentCaptured(
         bytes32 indexed paymentId,
         address indexed payer,
         address indexed payee,
-        uint256 amount,
+        uint120 amount,
         uint120 capturableAmount,
         uint120 refundableAmount
     );
@@ -1953,7 +1958,7 @@ contract RAIL0Test is Test {
         bytes32 indexed paymentId,
         address indexed payer,
         address indexed payee,
-        uint256 amount,
+        uint120 amount,
         uint120 capturableAmount,
         uint120 refundableAmount
     );
@@ -1961,7 +1966,7 @@ contract RAIL0Test is Test {
         bytes32 indexed paymentId,
         address indexed payer,
         address indexed payee,
-        uint256 amount,
+        uint120 amount,
         uint120 capturableAmount,
         uint120 refundableAmount
     );
@@ -1969,7 +1974,7 @@ contract RAIL0Test is Test {
         bytes32 indexed paymentId,
         address indexed payer,
         address indexed payee,
-        uint256 amount,
+        uint120 amount,
         uint120 capturableAmount,
         uint120 refundableAmount
     );
